Split Carousel's sub-components into their own files
