+++ frontend/components/carousel/aria-live/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { AriaLiveProps } from './types';
