+++ frontend/components/carousel/aria-live/types.ts
@@ -1,0 +1,6 @@
+export type AriaLiveProps = {
+  /** Total number of slides */
+  numSlides: number;
+  /** Current slide index */
+  currentSlide: number;
+};
