+++ frontend/components/carousel/arrows/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { ArrowsProps } from './types';
