+++ frontend/components/carousel/index.ts
@@ -1,2 +1,7 @@
-export { default, Slide } from './component';
+import Slide, { SlideProps } from './slide';
+
+export { default } from './component';
+export { Slide };
+
 export type { CarouselProps } from './types';
+export type { SlideProps };
