+++ frontend/components/carousel/paging/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { PagingProps } from './types';
