+++ frontend/components/carousel/slide/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { SlideProps } from './types';
